Create Installation Client (#15)
- Adds `NewInstallationClient` to the `internal/client` package,
allowing creation of GitHub App Installation Clients.
- Adds `GetInstallationClients` to the `AppsClient` process, allowing
creation of Installation Clients for all current App Installations.

Closes #12

+++ config/config.go
@@ -5,7 +5,7 @@ import (
 )
 
 type GitHubAppConfig struct {
-	AppId          int
+	AppId          int64
 	PrivateKeyPath string
 	WebhookSecret  string
 }
@@ -15,7 +15,7 @@ var gitHubAppConfig *GitHubAppConfig
 func GetGitHubAppConig() GitHubAppConfig {
 	if gitHubAppConfig == nil {
 		gitHubAppConfig = &GitHubAppConfig{
-			AppId:          viper.GetInt("APP_ID"),
+			AppId:          viper.GetInt64("APP_ID"),
 			PrivateKeyPath: viper.GetString("PRIVATE_KEY_PATH"),
 			WebhookSecret:  viper.GetString("WEBHOOK_SECRET"),
 		}
